# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a server for URL Shortener API.",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/shorted": {
            "post": {
                "description": "Create a shortened URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shorten"
                ],
                "summary": "Shorten URL",
                "operationId": "shorten-url",
                "parameters": [
                    {
                        "description": "Original URL",
                        "name": "shorten",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.shortenBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.message"
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "Redirect to original URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redirect"
                ],
                "summary": "Redirect to original URL",
                "operationId": "redirect-url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shortened URL",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "originalURL": {
                    "type": "string"
                },
                "shortenedURL": {
                    "type": "string"
                }
            }
        },
        "main.shortenBody": {
            "type": "object",
            "properties": {
                "original_url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}